Source Code: TFTP Client Implementation
This project is a robust implementation of a Trivial File Transfer Protocol (TFTP) client, developed in strict accordance with the packet rules outlined in the official RFC (Request for Comments). The client is designed to facilitate the seamless downloading of files from a remote host to a local target system.